refactor(cli): split show-error into focused modules + UX polish + Sentry observability (SHERLO-1173)
- Split monolithic showError.ts into focused submodules: detectPlatform, detectBundler/, fetchError, buildSourceMaps/, symbolicateFrames/, renderOutput
- UX: ora spinner, condensed 'Detected: <platform> · <projectType> · <engine>' line, framed error blocks (chalk.cyan), red.bold error name, '✓ Done' marker
- Sentry observability: errorToReport plumbed through 4 catch sites (no longer silent failures in production)
- Tests updated for new shape
