+++ packages/cli/src/commands/showError/buildSourceMaps/index.ts
@@ -1,0 +1,2 @@
+export { default } from './buildSourceMaps';
+export { default as findSourceMap } from './findSourceMap';
