+++ packages/cli/src/commands/showError/detectBundler/index.ts
@@ -1,0 +1,4 @@
+export { default } from './detectBundler';
+export { default as detectEntryFile } from './detectEntryFile';
+export { default as parseAndroidBundleCommand } from './parseAndroidBundleCommand';
+export { default as parseIosBundleCommand } from './parseIosBundleCommand';
