+++ packages/cli/src/commands/showError/symbolicateFrames/index.ts
@@ -1,0 +1,2 @@
+export { default } from './symbolicateAllFrames';
+export { default as runMetroSymbolicate } from './runMetroSymbolicate';
